refactor: 공고가 없는 경우와 있는 경우의 컴포넌트 분리(#197)
