+++ src/components/recruitment/recruitment-list/JobPostList.tsx
@@ -1,0 +1,53 @@
+import JobPostCard from '@components/commons/JobPostCard'
+import Button from '@components/commons/button/Button'
+import { Plus as PlusIcon } from 'lucide-react'
+import type { JobPost } from '@src/types/jobPosts'
+import type { RefObject } from 'react'
+
+interface JobPostListProps {
+  jobs: JobPost[]
+  infiniteMode: boolean
+  setInfiniteMode: (value: boolean) => void
+  loadMoreRef: RefObject<HTMLDivElement | null>
+  isFetchingNextPage: boolean
+  totalCount: number | undefined
+}
+
+export default function JobPostList({
+  jobs,
+  infiniteMode,
+  setInfiniteMode,
+  loadMoreRef,
+  isFetchingNextPage,
+  totalCount,
+}: JobPostListProps) {
+  return (
+    <div>
+      <h4 className="w-full pb-[24px] text-xl font-semibold">
+        전체 공고 ({totalCount})
+      </h4>
+      <ul className="w-[1216px]">
+        {jobs.map((job) => (
+          <li key={job.id} className="rounded-lg bg-white">
+            <JobPostCard post={job} />
+          </li>
+        ))}
+      </ul>
+
+      <div className="flex h-30 items-center justify-center">
+        {!infiniteMode ? (
+          <Button
+            buttonInnerText="더 많은 공고 보기"
+            icon={PlusIcon}
+            variant="secondary"
+            onClick={() => setInfiniteMode(true)}
+          />
+        ) : (
+          <div ref={loadMoreRef} style={{ height: '1px' }} />
+        )}
+
+        {isFetchingNextPage && <p className="text-gray-600">불러오는 중...</p>}
+      </div>
+    </div>
+  )
+}
